feat: refactor lexical mention plugin
