+++ apps/web/src/components/Composer/Editor/index.tsx
@@ -1,9 +1,9 @@
 import type { FC } from 'react';
 
-import MentionsPlugin from '@components/Shared/Lexical/Plugins/AtMentionsPlugin';
 import LexicalAutoLinkPlugin from '@components/Shared/Lexical/Plugins/AutoLinkPlugin';
 import EmojiPickerPlugin from '@components/Shared/Lexical/Plugins/EmojiPicker';
 import ImagesPlugin from '@components/Shared/Lexical/Plugins/ImagesPlugin';
+import MentionsPlugin from '@components/Shared/Lexical/Plugins/MentionsPlugin';
 import { Errors } from '@hey/data/errors';
 import getAvatar from '@hey/lib/getAvatar';
 import { Image } from '@hey/ui';
