+++ packages/data/regex.ts
@@ -4,7 +4,7 @@ export const Regex = {
   ethereumAddress: /^(0x)?[\da-f]{40}$/i,
   handle: /^[\dA-Za-z]\w{4,25}$/g,
   hashtag: /(#\w*[A-Za-z]\w*)/g,
-  mention: /@\w+\/[\w@]+/g,
+  mention: /@\w+\/([\dA-Za-z]\w{2,25})/g,
   profileNameFilter: new RegExp(`[${RESTRICTED_SYMBOLS}]`, 'gu'),
   profileNameValidator: new RegExp(`^[^${RESTRICTED_SYMBOLS}]+$`),
   // modified version of https://stackoverflow.com/a/6041965/961254 to support unicode international characters
